# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Seeds API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Current session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/applications:
    get:
      operationId: listApplications
      tags: [admin]
      summary: List applications (admin)
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
        - in: query
          name: applicationStatus
          schema:
            $ref: "#/components/schemas/ApplicationLifecycleStatus"
        - in: query
          name: finalDecision
          schema:
            $ref: "#/components/schemas/FinalDecision"
        - in: query
          name: evaluationCompletion
          schema:
            type: string
            enum: [any, none, partial, complete]
        - in: query
          name: interviewStatus
          schema:
            $ref: "#/components/schemas/InterviewStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/applications/{id}:
    get:
      operationId: getApplication
      tags: [admin]
      summary: Get application detail (admin) — includes assignments, evaluations, interview, decision logs
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateApplication
      tags: [admin]
      summary: Update application status / admin note (legacy MVP1)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    UserRole:
      type: string
      enum: [admin, evaluator]
    ApplicationStatus:
      type: string
      enum:
        - submitted
        - reviewing
        - interview
        - accepted
        - rejected
        - waitlisted
        - withdrawn
    ApplicationLifecycleStatus:
      type: string
      enum:
        - submitted
        - document_review
        - document_review_completed
        - interview
        - interview_scheduled
        - interview_completed
        - final_decision_made
        - withdrawn
    FinalDecision:
      type: string
      enum: [pending, accepted, rejected, waitlisted, withdrawn]
    EvaluationStage:
      type: string
      enum: [document_review, interview]
    AssignmentStatus:
      type: string
      enum: [assigned, in_progress, completed]
    InterviewStatus:
      type: string
      enum: [not_scheduled, scheduled, completed, no_show, cancelled]
    Recommendation:
      type: string
      enum: [strong_accept, accept, hold, reject, strong_reject]
...
    Application:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        school: { type: string }
        grade: { type: string }
        birthYear: { type: integer }
        interestArea: { type: string }
        motivation: { type: string }
        experience: { type: string }
        problemAwareness: { type: string }
        expectation: { type: string }
        privacyConsent: { type: boolean }
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        applicationStatus:
          $ref: "#/components/schemas/ApplicationLifecycleStatus"
        finalDecision:
          $ref: "#/components/schemas/FinalDecision"
        adminNote:
          type: string
          nullable: true
        submittedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - name
        - email
        - phone
        - school
        - grade
        - birthYear
        - interestArea
        - motivation
        - experience
        - problemAwareness
        - expectation
        - privacyConsent
        - status
        - applicationStatus
        - finalDecision
        - adminNote
        - submittedAt
        - updatedAt
    ApplicationListItem:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        school: { type: string }
        grade: { type: string }
        interestArea: { type: string }
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        applicationStatus:
          $ref: "#/components/schemas/ApplicationLifecycleStatus"
        finalDecision:
          $ref: "#/components/schemas/FinalDecision"
        avgDocReviewScore:
          type: number
          nullable: true
        evaluationsAssigned: { type: integer }
        evaluationsCompleted: { type: integer }
        interviewStatus:
          $ref: "#/components/schemas/InterviewStatus"
        submittedAt: { type: string, format: date-time }
      required:
        - id
        - name
        - email
        - school
        - grade
        - interestArea
        - status
        - applicationStatus
        - finalDecision
        - avgDocReviewScore
        - evaluationsAssigned
        - evaluationsCompleted
        - interviewStatus
        - submittedAt
    ApplicationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationListItem"
        total: { type: integer }
      required: [items, total]
...
    UpdateApplicationBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        applicationStatus:
          $ref: "#/components/schemas/ApplicationLifecycleStatus"
        adminNote:
          type: string
          nullable: true
          maxLength: 5000
    AdminLoginBody:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
      required: [email, password]
    SessionUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
      required: [id, email, name, role]
...
    EvaluationAssignment:
      type: object
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        evaluatorId: { type: integer }
        evaluatorName: { type: string }
        evaluatorEmail: { type: string }
        stage:
          $ref: "#/components/schemas/EvaluationStage"
        status:
          $ref: "#/components/schemas/AssignmentStatus"
        assignedAt: { type: string, format: date-time }
      required: [id, applicationId, evaluatorId, evaluatorName, evaluatorEmail, stage, status, assignedAt]
...
    Evaluation:
      type: object
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        evaluatorId: { type: integer }
        evaluatorName: { type: string }
        stage:
          $ref: "#/components/schemas/EvaluationStage"
        motivationScore: { type: integer, nullable: true }
        problemAwarenessScore: { type: integer, nullable: true }
        initiativeScore: { type: integer, nullable: true }
        collaborationScore: { type: integer, nullable: true }
        fitScore: { type: integer, nullable: true }
        overallScore: { type: integer }
        recommendation:
          $ref: "#/components/schemas/Recommendation"
        comment:
          type: string
          nullable: true
        submittedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - applicationId
        - evaluatorId
        - evaluatorName
        - stage
        - motivationScore
        - problemAwarenessScore
        - initiativeScore
        - collaborationScore
        - fitScore
        - overallScore
        - recommendation
        - comment
        - submittedAt
        - updatedAt
...
    Interview:
      type: object
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        locationOrLink:
          type: string
          nullable: true
        interviewerNote:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/InterviewStatus"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, applicationId, scheduledAt, locationOrLink, interviewerNote, status, createdAt, updatedAt]
...
    DecisionLog:
      type: object
      properties:
        id: { type: integer }
        applicationId: { type: integer }
        previousDecision:
          allOf:
            - $ref: "#/components/schemas/FinalDecision"
          nullable: true
        newDecision:
          $ref: "#/components/schemas/FinalDecision"
        changedBy:
          type: integer
          nullable: true
        changedByName:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
      required: [id, applicationId, previousDecision, newDecision, changedBy, changedByName, reason, createdAt]
...
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: "#/components/schemas/EvaluationAssignment"
            evaluations:
              type: array
              items:
                $ref: "#/components/schemas/Evaluation"
            interview:
              oneOf:
                - $ref: "#/components/schemas/Interview"
                - type: "null"
            decisionLogs:
              type: array
              items:
                $ref: "#/components/schemas/DecisionLog"
            avgDocReviewScore:
              type: number
              nullable: true
          required: [assignments, evaluations, interview, decisionLogs, avgDocReviewScore]
...