# lib/api-zod/src/generated/types/applicationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Seeds API
 * OpenAPI spec version: 0.2.0
 */
import type { Application } from "./application";
import type { DecisionLog } from "./decisionLog";
import type { Evaluation } from "./evaluation";
import type { EvaluationAssignment } from "./evaluationAssignment";
import type { Interview } from "./interview";

export type ApplicationDetail = Application & {
  assignments: EvaluationAssignment[];
  evaluations: Evaluation[];
  interview: Interview | null;
  decisionLogs: DecisionLog[];
  avgDocReviewScore: number | null;
};