# lib/api-zod/src/generated/types/applicationList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Seeds API
 * OpenAPI spec version: 0.2.0
 */
import type { ApplicationListItem } from "./applicationListItem";

export interface ApplicationList {
  items: ApplicationListItem[];
  total: number;
}