# lib/api-zod/src/generated/types/updateApplicationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Seeds API
 * OpenAPI spec version: 0.2.0
 */
import type { ApplicationLifecycleStatus } from "./applicationLifecycleStatus";
import type { ApplicationStatus } from "./applicationStatus";

export interface UpdateApplicationBody {
  status?: ApplicationStatus;
  applicationStatus?: ApplicationLifecycleStatus;
  /** @maxLength 5000 */
  adminNote?: string | null;
}